[UpSkill-Dev][Q1-2023-24]
- Fetched monthly records by user
- Added up the monthly sum
- Passed the addition result to the predictor method.

+++ Monthly Expenditure Predictor/Models/ExpenditurePredictionParams.swift
@@ -13,7 +13,7 @@ class ExpenditurePredictionParams{
          address: String,
          noOfFamilyMember: Int,
          expenseAmount: Int,
-         cumulativeSum: Int,
+         cumulativeSum: Double,
          month: Int,
          day: Int) 
     {
